crash_recovery/pyproject.toml
[project]
name = "summarize"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Cully Wakelin", email = "cully.wakelin@resonatehq.io" }
]
dependencies = [
    "resonate-sdk==0.4.12",
    "flask==3.0.3",
]
readme = "README.md"
requires-python = "== 3.12.*"

[project.scripts]
    "app" = "src.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]
